# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation",
        "description": "Documentation for the Car Reservation API",
        "contact": {
            "email": "admin@example.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/jadwal_dokter": {
            "get": {
                "tags": [
                    "Jadwal Dokter"
                ],
                "summary": "Tampilkan semua jadwal dokter",
                "operationId": "a4e6a7bd4e06ee80009674209e0b0ec3",
                "responses": {
                    "200": {
                        "description": "Berhasil menampilkan semua jadwal dokter"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Jadwal Dokter"
                ],
                "summary": "Tambah jadwal dokter baru",
                "operationId": "c89707ed78a095a1ab1ad1c5422b629a",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "dokter_id",
                                    "hari",
                                    "jam_mulai",
                                    "jam_selesai"
                                ],
                                "properties": {
                                    "dokter_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "hari": {
                                        "type": "string",
                                        "example": "Senin"
                                    },
                                    "jam_mulai": {
                                        "type": "string",
                                        "format": "time",
                                        "example": "08:00"
                                    },
                                    "jam_selesai": {
                                        "type": "string",
                                        "format": "time",
                                        "example": "12:00"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Jadwal berhasil ditambahkan"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/jadwal_dokter/{id}": {
            "get": {
                "tags": [
                    "Jadwal Dokter"
                ],
                "summary": "Tampilkan jadwal dokter berdasarkan ID",
                "operationId": "5be25cea8725e706c4d8aff1c6ce079a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Berhasil menampilkan jadwal dokter"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Jadwal Dokter"
                ],
                "summary": "Perbarui jadwal dokter",
                "operationId": "75dfd0083955c91bfeb30e39c17b197b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "dokter_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "hari": {
                                        "type": "string",
                                        "example": "Selasa"
                                    },
                                    "jam_mulai": {
                                        "type": "string",
                                        "format": "time",
                                        "example": "09:00"
                                    },
                                    "jam_selesai": {
                                        "type": "string",
                                        "format": "time",
                                        "example": "13:00"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Jadwal dokter berhasil diperbarui"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Jadwal Dokter"
                ],
                "summary": "Hapus jadwal dokter",
                "operationId": "8055070ecef2ac156f2e8e97d9ac8ff5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Jadwal dokter berhasil dihapus"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/jadwal_dokter_with_dokter": {
            "get": {
                "tags": [
                    "Jadwal Dokter"
                ],
                "summary": "Tampilkan semua jadwal dokter beserta data dokternya",
                "operationId": "546e4ff651e027b735904472f11e297e",
                "responses": {
                    "200": {
                        "description": "Berhasil menampilkan data jadwal dokter dengan relasi dokter",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Data jadwal dokter ditemukan"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "hari": {
                                                        "type": "string",
                                                        "example": "Senin"
                                                    },
                                                    "jam_mulai": {
                                                        "type": "string",
                                                        "example": "08:00"
                                                    },
                                                    "jam_selesai": {
                                                        "type": "string",
                                                        "example": "12:00"
                                                    },
                                                    "dokter": {
                                                        "properties": {
                                                            "id": {
                                                                "type": "integer",
                                                                "example": 1
                                                            },
                                                            "nama": {
                                                                "type": "string",
                                                                "example": "dr. Andi"
                                                            },
                                                            "spesialis": {
                                                                "type": "string",
                                                                "example": "Umum"
                                                            },
                                                            "no_telp": {
                                                                "type": "string",
                                                                "example": "08123456789"
                                                            },
                                                            "email": {
                                                                "type": "string",
                                                                "example": "andi@example.com"
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/jadwal_dokter_Hari_Ini": {
            "get": {
                "tags": [
                    "Jadwal Dokter"
                ],
                "summary": "Menampilkan daftar dokter yang tersedia berdasarkan hari",
                "description": "Endpoint ini menampilkan daftar dokter yang siap praktik berdasarkan hari. Jika tidak diberikan parameter hari, maka akan menggunakan hari ini.",
                "operationId": "081d1e602f9973c2ac16eca4a6d3c15e",
                "parameters": [
                    {
                        "name": "hari",
                        "in": "query",
                        "description": "Hari dalam format string (contoh: Senin, Selasa, Rabu, dst). Jika tidak diisi, otomatis ambil hari ini.",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "Senin"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Daftar dokter tersedia",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "nama": {
                                                "type": "string",
                                                "example": "dr. Sinta Dewi"
                                            },
                                            "spesialis": {
                                                "type": "string",
                                                "example": "Spesialis Anak"
                                            },
                                            "hari": {
                                                "type": "string",
                                                "example": "Senin"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/jadwal_dokter/search": {
            "get": {
                "tags": [
                    "Jadwal Dokter"
                ],
                "summary": "Cari jadwal dokter berdasarkan nama dokter (tanpa autentikasi)",
                "operationId": "3b95ec0095b3f0764898bb989e06c30d",
                "parameters": [
                    {
                        "name": "nama",
                        "in": "query",
                        "description": "Nama dokter yang ingin dicari (bisa sebagian)",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "Andi"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Data jadwal dokter ditemukan"
                    },
                    "404": {
                        "description": "Data tidak ditemukan"
                    }
                }
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "description": "Masukkan token dalam format: Bearer {token}"
            }
        }
    },
...
    "security": [
        {
            "bearerAuth": []
        }
    ]
}